% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fib.R
\name{fib_before_n}
\alias{fib_before_n}
\title{Returns fibonacci numbers up to a given integer}
\usage{
fib_before_n(n)
}
\arguments{
\item{n}{Integer giving the upper bound}
}
\value{
Integer vector containing numbers less than n
}
\description{
Returns fibonacci numbers up to a given integer
}
\examples{
fib_before_n(10)
[1] 0 1 1 2 3 5 8
}
